// src/app/core/shared/models/mssql/IDxcSearchReasonItem.ts
export interface IDxcSearchReasonItem {
    id: string;
    personal: string;
    caseInfo: string;
    reasons: string;
    agree: string;
    authenId: string;
    userId: string;
    timestamp: string;
    reasons_select: string;

    _dateOfExpire:number // จำนวนวันที่เหลือให้ใช้งาน
}